+++ src/algorithms/InsertionSort.java
@@ -4,9 +4,9 @@ import java.util.LinkedList;
 
 public class InsertionSort implements  SortingAlgorithm{
 
-    private LinkedList<Integer> insertionSortHelper(LinkedList<Integer> list){
+    private LinkedList<Long> insertionSortHelper(LinkedList<Long> list){
         for(int j = 2;j <  list.size(); j++){
-            int chave = list.get(j);
+            long chave = list.get(j);
             int i = j -1;
             while(i >= 0 && list.get(i) > chave){
                 list.set(i + 1,list.get(i));
@@ -17,7 +17,7 @@ public class InsertionSort implements  SortingAlgorithm{
         return list;
     }
     @Override
-    public LinkedList<Integer> sort(LinkedList<Integer> list) {
+    public LinkedList<Long> sort(LinkedList<Long> list) {
         return insertionSortHelper(list);
     }
 }
