+++ src/algorithms/SelectionSort.java
@@ -3,15 +3,15 @@ package src.algorithms;
 import java.util.LinkedList;
 
 public class SelectionSort implements SortingAlgorithm{
-    public LinkedList<Integer> slectionHelper(LinkedList<Integer> list){
+    public LinkedList<Long> slectionHelper(LinkedList<Long> list){
             int lowestIndex;
             for(int i = 0; i < list.size(); i ++){
                 for(int j = i; j < list.size(); j++){
-                    int currentLowest = j;
+                    long currentLowest = j;
                     if(list.get(i) > list.get(j + 1) && currentLowest > list.get(j + 1)){
                         currentLowest = list.get(j + 1);
                     }else{
-                        int swap = list.get(i);
+                        long swap = list.get(i);
                         list.set(i, list.get(j+1));
                         list.set(j, swap);
                     }
@@ -24,7 +24,7 @@ public class SelectionSort implements SortingAlgorithm{
     }
 
     @Override
-    public LinkedList<Integer> sort(LinkedList<Integer> lista) {
+    public LinkedList<Long> sort(LinkedList<Long> lista) {
         return null;
     }
 }
